chore: Document important information on MMC support
This documents inportant information on MMC support in MMC-SD for MMC Linux support
and UG-Memory-K3 for MMC u-boot support.

This includes a refactor of the existing instance that documented errata i2458 for
eMMC, and making the instance more generalistic to be able to reuse for all MMC related
support updates.

Signed-off-by: Judith Mendez <jm@ti.com>